Define local vars "spawnx", "spawnz", "spawna", and "spawnat" in spawnscript.
"spawnx", "spawnz", and "spawna" are the coordinates defined by the "coords"
command, and "spawnat" is the scroll position defined by "at" in the level
file.

Useful for comparing the intended spawn coordinates with the actual spawn
location, which may be different if, for example, the scroll position has
been changed by script.