Move the deprecation warning for the "a" property to the mapstrings function.
This way, the warning is written to the log only once (at compile time) for
each use of the "a" property, rather than at run time.  Doing it at run time
caused major I/O-bound slowdowns and an absurdly large log file if a mod
accessed the property in an update script.

Also fixes the spelling of "deprecation" in the warning message.