Don't create directory structure for Vita release directory
The Paks, Saves, etc. directories have to be at a fixed path on the
Vita (ux0:/data/OpenBOR/), and aren't relative to the OpenBOR package or
executable. No need to confuse the user by implying that they need to copy
the empty directories over to their Vita when OpenBOR will create them on
startup anyway.