Player Spawn Optimizing
*Player spawn is now a dynamically allocated pointer so we can work with
it as a collection in script. Also saves a little memory on modules with
less than four players since we only allocate space for the current
maximum number of players.
*Default spawn Y position for players is now the current vertical
resolution + 60 instead of a static 300. This keeps a consistent default
behavior in higher res modules without the author having to manually
compensate.