Fix reference counting bug introduced by b6e3f4131
This is the fix for the underlying cause of the NULL dereference mentioned
in 9e690e6fe. There was a case where we were calling StrCache_Collect() on
an uninitialized string variant. Add asserts to guard against this in the
future, and initialize vt->strVal to an invalid value in
ScriptVariant_ChangeType() since 0 is a valid string cache index.