Handle Based Animation Property In progress
Created a branch so I could work and track progress on handle based
animation property project without breaking the master or crossing wires
with other commits. Created file for coordinates - it's just a copt
paste job of body_collision.c at the moment, but will be focus of next
update.