Deprciate altbase.
On further inspection of 958d892, seta is in fact used. However, the entity member altbase is superfluous. It is intended for movea, but movea just sets the base directly anyway, and removal of altbase will allow better internal consistency. I am trying to get a uniform set of axis structs and have anything in the engine using axes to use these without custom members.