Fix input mappings being assigned to the wrong controller
In SDL 2, each joystick has a "device index" and an "instance ID", and at
least on Windows, they can be different numbers. We were trying to use
them interchangeably, and the result was that the control config menu
would disagree with the rest of the engine about which controller was
which.

Fixes https://github.com/DCurrent/openbor/issues/70