Speed up the clipbitmap() function
The clipbitmap() function was actually moving pixels around in the image
when it really only needs to figure out which parts should be clipped,
and let encodesprite() encode only those parts of the image.

This improves load times slightly, since clipbitmap() was the biggest user
of CPU time when loading, but loading is heavily I/O bound, so CPU time
isn't a huge deal.