Fix cropped PNG decoding.
Fixes https://github.com/DCurrent/openbor/issues/155

The allocscreen() function in source/gamelib/screen.c truncates screen
widths to multiples of 4. In the linked issue, the decoder needed to
decode a 961*481 PNG to a 960*481 screen, cropping off the rightmost
column of pixels. But the new PNG decoder treated the cropped width as
the original image width, and thus misjudged the position of the filter
type bytes. Fix this by properly distinguishing between the original
(input) and cropped (output) image sizes.

There are still problems loading an interlaced 961*481 PNG as the title
screen background, but that's a pain to fix, and the linked issue only
concerns non-interlaced PNGs.