update pre-built libraries, and set to build android for both armeabi-v7a and arm64-v8a
Note this in turn updated both its binary library, and its corresponding
header files as well.

The specific versions of all dependencies are listed in README.md for both
armeabi-v7a/ and arm64-v8a/ directory for reference in keeping tracking
of.

All the dependencies are built against Android SDK with API level of 19
as minimum target SDK, as this allows accomodation in building
autotools-based library.

With additional of arm64-v8a, this relaxes the requirement of having
64-bit app on Play Store which will be enforced in August 2019.