add modified vibration logic for android via jni
Vibration logic is originally created by White Dragon. This changes
migrated it from previously being there in SDLActivity.java with
modification in the following 3 aspects

- use JNI (calling from C to Java): checking logic is in C, but
vibration logic is in Java. This reduces 1 round-trip time unnecessary
in switching back and forth if in that frame we don't need to do any
vibration
- migrate vibration code in Android as there is deprecation happened
- properly tune the vibrate pattern to be shorter (anyway, actually we
might not need to vibrate for every touch on every UI elements, leave
this for later)

For Android JNI part, look into jniutils[.cpp/.h], and removed old
commented un-used code.