make sure all deps for Android included debug-info
All those deps in openbor/lib/arm* directories are in total of 45 MB. For now, included them in-source
with pre-built ones (consider later in the future to include those deps' source in git index, then
allow users to build all from source...).

Note also those deps included with debug-info as gradle will strip debugging info automatically when
creating an apk file, and for us to reduce the complexity for now to not include multiple set of deps
i.e. debug (unoptimized + debug-info), and release (optimized + debug-info). That multiple set of deps
option is good to have but just for now doesn't useful for almost all OpenBOR users, and likely even
power users won't need them now. But anyway having 'debug' set will increase the risk of users not
testing 'release' set enough to found the bug of the game (if any). Reason is these two sets are benefit
mostly for developers adding additional features to android build of the game ontop of OpenBOR, but not
immediate users who mostly mod the game.

Ok, in short there are room for improvement to note, just not good time for now to add it.
1. Build all those deps from source, thus include deps' source in the repo
2. Having multiple set of profile deps i.e. debug, release