Android updates & increased fonts (#203)
* Android updates & increased fonts

Just updated the process for setting up the andoid sdk as the android command line tools have changed their folder structure.  Small updated to the packfile.c as it was breaking the android build. Updated the font limit from 10 to 100.  Updated the android readme on how to compile the android build.

* Reverted Font.h

Reverted font.h to allow DC to update.