Fix button names in the OpenBOR menu
This fix changes the detected keys in the OpenBOR menu by direct strings to avoid the "none" key names due to the last controller fix.

This change will not make a difference in the OpenBOR because the menu always uses the default engine keys, not the keys configured in the game.

So, the key names can be defined by strings in Windows, same as the Android or other platforms do.

NOTE: This change affects only the OpenBOR menu in the mods selector.