Reversed the FPS limit to reduce CPU usage
The 200 FPS in the previous builds were a waste of energy, and it's true.

However, by removing the FPS limit and when the player disable the V-SYNC, the FPS reaches a very high numbers like 3000+ depending on the hardware used, and it causes a bigger waste of energy than the 200 FPS, plus it causes a high CPU usage too (something between 30%-40%).

In this fix we can have the best of the two worlds:
- the Plombo's V-SYNC code intact, if enabled it runs at 60 FPS
- the previous code if the V-SYNC is disabled running at 200 FPS max, no more than that