Fixed the "edge screen stuck" bug
There's a long time bug since from the build 4432 than randomly stucks the entities at the edges of the screen.

To fix that I reduced the "screen checking" range from 10 to 5 to avoid the entities to stuck in the edge of the screen when the engine uses the "testmove" function.

This change was made because the "common_trymove" function also has another "screen checking" with a range of 10 too.

If the "testmove" function has a equal or bigger range than the "common_trymove" function, sometimes the entities will stuck.