Changed some button tasks priorities in the select screen
In the select screen, if you press "attack + right or left" buttons together when a character is highlighted, sometimes he will not be confirmed. It happens because the code that confirms the character is placed after the code that highlights and changes him. So, the engine will prioritize the "character change" task over the "character confirmation" task.

To fix this problem, I moved the "confirmation" code to a place before the "highlight" code. This way, a character will always be confirmed when the "attack + right or left" buttons are pressed together because the "character confirmation" task will always come before all others.