October-2021 Updates
Translation
- Added an external location for translation files, outside of the PAK file in the folder "Translation/".
- Extended translation for level sets.
- Made the translation function work during the "menu screen" to allow translation "on-the-fly".- Moved the "menu" and "fonts" functions closer to the translation function during the OpenBOR startup.- Small addition of a second instance of the "control_init" function to help translate control config menu.

Blocking
- Fixed the "Blocking" instance during "non-blocking" animations (pain, walk, idle, etc).
- Fixed the "Blockodds" function to work properly according to the value defined on the entities.
- Fixed the "Toflip" function, now if the flag is 0 the "flash" will get the same direction as the defender.

Cheats
- Fixed some "Infinite health" cheat bugs ("in-hole" and "special attack" bugs).

Grabs and Vault
- Made grabflip property accessible by script.- Added new backgrab animation.
- Added new vault animation.
- Added new vault2 animation.

Special and Jumping attacks
- Added new "Ajspecial" flags to have more native "special attack" keys (flags 2,3 and 4 corresponds the a2, a3 and a4 keys).
- Added the new "jumpspecial" property that allows the modder to enable or disable the default animation movement.

By default, this animation "kills" the "xyz" movement while in the air, and now if this new property is used with the flag 1, the movement will not be killed, working the same as any jumping attack.

The native behaviour is maintained, old games don't need to change anything, this is useful for new games if the modder wants to use the "special3/jumpspecial" animation in a different way.

Can be edited by declaring at the entity header or by script using the "entity property".

- Fixed a bug in the "jumpspecial" that constantly consumes health or mp if the special button is constantly pressed, even if the current "jumpspecial" animation cycle is not finished yet.
- Fixed the noaircancel function bug, now the flag "2" works as intended and completely disables the cancelation between all jumping attacks.

In the previous code, both flags 1 and 2 have the same effect and allow the cancelation after last "jumpattack" is finished.

- Made the "noaircancel" function accessible by script using "openborvariant".

For backward compatibility, the models.txt native function works normally, but the scripted version will override the models.txt function if changed.
To edit it, you need to use "changeopenborvariant("noaircancel", value);".

NOTES:

The backgrab animation follows a similar form of other animations.
Ex: walk/backwalk, grab/backgrab.

The vault animations follow the same format and rules of all grab attacks.
Ex: grabup/grabup2, grabdown/grabdown2, vault/vault2.

The first vault can be automatically performed by pressing the jump button when an opponent is grabbed, and of course only when the animation is declared in the entity's file. Otherwise, the grabattack2 or attack3 will be performed by default.

The vault2 can only be used with scripts for now, but it's important to have because there's some differences when an opponent is grabbed from behind.

This new code maintains all previous functionality for backward compatibility. Vault animations don't have any predefined instance yet (aiflag, takeaction, etc), they are "empty" same as follow animations.

Your focus is to give a proper native functionality for an important mechanic present in the entire SOR franchise (back grab and vault animations) and avoid losing some "follows" or "grabattacks" to make it happen.

When the jump button is pressed, the first vault animation will be prioritized over grabattack2 and attack3 animations, while grabbing an opponent. If there's no vault animation, the order is grabattack2 and then attack3. It means that the vault won't get in the way of the default system if it's not declared and will only be used when the modder knows what he is doing.